Binance/backend/requirements.txt
python-binance
Flask
pymongo
python-dotenv